requirements.txt
streamlit
openai
langchain
langchain-community
faiss-cpu
tiktoken
python-dotenv
pypdf